requirements.txt
Flask==3.0.3
python-docx==1.1.2
PyMuPDF==1.24.9
pillow==10.4.0
gunicorn==20.1.0
